Add Accounting feature to depository and controller (#53)

+++ programs/uxd/src/instructions/set_redeemable_global_supply_cap.rs
@@ -22,7 +22,7 @@ pub struct SetRedeemableGlobalSupplyCap<'info> {
 
 pub fn handler(
     ctx: Context<SetRedeemableGlobalSupplyCap>,
-    redeemable_global_supply_cap: u64, // In UI amount
+    redeemable_global_supply_cap: u128, // In Redeemable Native amount
 ) -> ProgramResult {
     ctx.accounts.controller.redeemable_global_supply_cap = redeemable_global_supply_cap;
     Ok(())
